refactor on First and Second common consent steps, splitted into new AboutTheStudy and SummaryAndSignature sections
