+++ src/components/dashboard/Dashboard.tsx
@@ -17,9 +17,7 @@ export const Dashboard: React.FunctionComponent<DashboardProps> = ({
 }: DashboardProps) => {
   const [error, setError] = useState()
   const [isLoading, setIsLoading] = useState(false)
-  const [userInfo, setUserInfo] = useState<LoggedInUserData | undefined>(
-    undefined,
-  )
+  const [userInfo, setUserInfo] = useState<LoggedInUserData | undefined>()
   const { push } = useHistory()
   const { t } = useTranslation()
 
@@ -33,11 +31,7 @@ export const Dashboard: React.FunctionComponent<DashboardProps> = ({
         try {
           setIsLoading(true)
           const userInfoResponse = (await UserService.getUserInfo(token)) as any
-          if (
-            userInfoResponse?.data &&
-            userInfoResponse?.data?.clientData?.consented &&
-            userInfoResponse?.data?.clientData?.skipRanking
-          ) {
+          if (userInfoResponse?.data?.clientData?.consented) {
             push(ROUTES.DOWNLOAD)
           }
